+++ src/simmate/apps/ppdb/components/ppdb_molecule_table.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from simmate.website.htmx.components import DynamicTableForm, MoleculeInput
+from simmate.website.data_explorer.components import DynamicTableForm
+from simmate.website.htmx.components import MoleculeInput
 
 from ..models import PpdbMolecule