+++ src/simmate/website/data_explorer/components/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .dynamic_table_form import DynamicTableForm